contrib: make EmbedLayerNormalization CUDA output write index 64-bit
For large batch * sequence_length * hidden_size (which can exceed 2^31
elements on high-memory GPUs), the EmbedLayerNormalization output write
index computed as sequence_position * hidden_size overflowed 32-bit
arithmetic and wrapped negative, producing an out-of-range write.

Make the whole output write-index path 64-bit end-to-end:
- Pass the already-widened int64 output offset straight to LayerNorm
  instead of narrowing it back to int.
- Widen the shared LayerNorm helper's offset parameter and its internal
  index from int to int64_t so the offset is not truncated again. This is
  source-compatible with the only other caller (skip layer norm passes an
  int offset, which promotes cleanly).

+++ onnxruntime/contrib_ops/cuda/bert/embed_layer_norm_impl.cu
@@ -197,8 +197,8 @@ __global__ void EmbedLayerNormKernel(
     thread_data = pair_sum(thread_data, cub::KeyValuePair<float, float>(rldval, rldval * val_f));
   }
 
-  // 3. layer norm on the sum (LayerNorm takes an int offset; the widened offset fits the output)
-  LayerNorm<T, TPB>(thread_data, hidden_size, static_cast<int>(output_offset), beta, gamma, epsilon, output);
+  // 3. layer norm on the sum (offset is int64 end-to-end to keep the output write index in range)
+  LayerNorm<T, TPB>(thread_data, hidden_size, output_offset, beta, gamma, epsilon, output);
 }
 
 template <typename T>
